+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog Fab-manager
 
+## v5.2.0 2021 December 23
+
+- Ability to configure prices for spaces by time slots different than the default hourly rate
+- Updated portuguese translation
+- Refactored the ReserveButton component to use the same user's data across all the component
+- First optimization the load time of the payment schedules list
+
 ## v5.1.13 2021 November 16
 
 - Fix a bug: unable to run the setup/upgrade scripts as root